+++ server/src/main/java/com/adobe/testing/s3mock/dto/Retention.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright 2017-2024 Adobe.
+ *  Copyright 2017-2025 Adobe.
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 
 package com.adobe.testing.s3mock.dto;
 
+import com.adobe.testing.S3Verified;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -28,17 +29,15 @@ import java.time.Instant;
  * For unknown reasons, the timestamps in the Retention are serialized in Nanoseconds instead of
  * Milliseconds, like everywhere else.
  */
+@S3Verified(year = 2025)
 @JsonRootName("Retention")
 public record Retention(
-    @JsonProperty("Mode")
-    Mode mode,
-    @JsonProperty("RetainUntilDate")
+    @JsonProperty("Mode") Mode mode,
     @JsonSerialize(using = InstantSerializer.class)
     @JsonDeserialize(using = InstantDeserializer.class)
-    Instant retainUntilDate,
+    @JsonProperty("RetainUntilDate") Instant retainUntilDate,
     //workaround for adding xmlns attribute to root element only.
-    @JacksonXmlProperty(isAttribute = true, localName = "xmlns")
-    String xmlns
+    @JacksonXmlProperty(isAttribute = true, localName = "xmlns") String xmlns
 ) {
   public Retention {
     if (xmlns == null) {